+++ src/hooks/queries/index.ts
@@ -47,6 +47,12 @@ export {
 } from './useNotificationsQuery';
 
 export {
+  notificationPreferencesKeys,
+  useNotificationPreferencesQuery,
+  useUpdatePreferenceMutation,
+} from './useNotificationPreferencesQuery';
+
+export {
   pharmacyKeys,
   usePharmacyCategoriesQuery,
   useFeaturedDrugsQuery,